# lib/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DeckAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectBodyDensity } from "./createProjectBodyDensity";

export interface CreateProjectBody {
  name: string;
  description?: string;
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  headingFont?: string;
  bodyFont?: string;
  density?: CreateProjectBodyDensity;
}