# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DeckAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyDensity } from "./updateProjectBodyDensity";

export interface UpdateProjectBody {
  name?: string;
  description?: string;
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  headingFont?: string;
  bodyFont?: string;
  logoObjectPath?: string | null;
  density?: UpdateProjectBodyDensity;
}